feat: Break apart the large file more
https://harperdb.atlassian.net/browse/STUDIO-483
